--- WolvenKit.Common/Tools/JsonConverters.cs
@@ -1,394 +1,0 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Linq;
-using System.Numerics;
-using System.Runtime.Serialization;
-using Newtonsoft.Json.Linq;
-using WolvenKit.Common.Model;
-using WolvenKit.Common.Model.Cr2w;
-using WolvenKit.Interfaces.Core;
-using WolvenKit.RED4.CR2W;
-using WolvenKit.RED4.CR2W.Reflection;
-using WolvenKit.RED4.CR2W.Types;
-
-namespace WolvenKit.Common.Tools
-{
-    public static class Red4Serializer
-    {
-        /// <summary>
-        /// Parse
-        /// </summary>
-        /// <param name="cVariable"></param>
-        /// <param name="value">either a JArray, JObject or primitive value</param>
-        /// <exception cref="InvalidParsingException"></exception>
-        public static void SetFromJObject(this IEditableVariable cVariable, object value)
-        {
-            // switch over the jobject
-            switch (value)
-            {
-                // enums and arrays are JArrays and can be directly deserialized
-                case JArray jArray:
-                    switch (cVariable)
-                    {
-                        // enums are serialized as list of strings
-                        case IREDEnum:
-                            var enumobj = jArray.ToObject<List<string>>();
-                            if (enumobj == null)
-                            {
-                                throw new InvalidParsingException("not a CVariable");
-                            }
-                            cVariable.SetValue(enumobj);
-                            break;
-                        // arrays
-                        case ICurveDataAccessor:
-                            throw new NotImplementedException();
-                            //break;
-                        case IREDArray redArray:
-                        {
-                            if (redArray.IsByteArray())
-                            {
-                                var bytes = jArray.ToObject<byte[]>();
-                                redArray.SetValue(bytes);
-                            }
-                            else
-                            {
-                                var listOfObjects = jArray.ToObject<List<object>>();
-                                if (listOfObjects == null)
-                                {
-                                    throw new InvalidParsingException("not a CVariable");
-                                }
-                                for (var i = 0; i < listOfObjects.Count; i++)
-                                {
-                                    var jitem = listOfObjects[i];
-
-                                    var element = redArray.GetElementInstance(i.ToString());
-                                    // parse the elements according to the array type
-                                    element.SetFromJObject(jitem);
-                                    redArray.AddVariable(element);
-                                }
-                            }
-
-                            break;
-                        }
-                        default:
-                            throw new InvalidParsingException("not a CVariable");
-                    }
-                    break;
-                // complex objects are JObjects and can be deserialized as CVariableDto (recursive)
-                case JObject jObject:
-                    cVariable.SetFromDictionary(jObject.ToObject<Dictionary<string,object>>());
-                    break;
-                // does that ever happen?
-                case JToken:
-                    throw new InvalidParsingException("file format not supported");
-                // primitive types can be set directly
-                default:
-                    switch (value)
-                    {
-                        // all numeric values
-                        case long:
-                        case double:
-                        case BigInteger:
-                            cVariable.SetValue(value.ToString()); // cast to string to do the parsing in the classes :/
-                            break;
-                        // other primitive types can be set directly
-                        case bool b:
-                            cVariable.SetValue(b);
-                            break;
-                        case string s:
-                            cVariable.SetValue(s);
-                            break;
-                        default:
-                            throw new InvalidParsingException("file format not supported");
-                    }
-
-                    break;
-            }
-        }
-
-        public static void SetFromDictionary(this IEditableVariable cvar, Dictionary<string,object> dictionary)
-        {
-            if (cvar is IREDVariant var)
-            {
-                var type = dictionary["Type"] as string;
-                var name = "Variant";
-                if (dictionary.ContainsKey("Name"))
-                {
-                    name = dictionary["Name"] as string;
-                }
-
-                var jvalue = dictionary["Variant"];
-                var variant = CR2WTypeManager.Create(type, name, cvar.Cr2wFile as IRed4EngineFile, cvar as CVariable);
-                variant.IsSerialized = true;
-                variant.SetFromJObject(jvalue);
-                var.SetVariant(variant);
-
-                return;
-            }
-
-            foreach (var (propertyName, value) in dictionary)
-            {
-                var redProperty = GetRedProperty(cvar, propertyName);
-                if (redProperty == null)
-                {
-                    throw new InvalidParsingException("not a CVariable");
-                }
-                redProperty.SetFromJObject(value);
-            }
-        }
-
-        private static CVariable GetRedProperty(IEditableVariable cvar, string propertyName)
-        {
-            foreach (var member in REDReflection.GetMembers(cvar))
-            {
-                try
-                {
-                    var redname = REDReflection.GetREDNameString(member);
-                    if (redname != propertyName)
-                    {
-                        continue;
-                    }
-
-                    var prop = cvar.accessor[cvar, member.Name];
-                    var cprop = prop as CVariable;
-
-                    return cprop;
-                }
-                catch (Exception e)
-                {
-                    Console.WriteLine(e);
-                    throw;
-                }
-            }
-
-            return null;
-        }
-
-        public static object ToObject(this IEditableVariable data)
-        {
-            switch (data)
-                {
-                    case CDateTime d:
-                        return d.ToUInt64();
-                    // special cases of primitive types
-                    case CVariant var:
-                        return new Dictionary<string, object>()
-                        {
-                            {"Type", var.Variant.REDType},
-                            {"Variant", var.Variant.ToObject()}
-                        };
-                    case CVariantSizeNameType varnt:
-                        return new Dictionary<string, object>()
-                        {
-                            {"Type", varnt.Variant.REDType},
-                            {"Name", varnt.Variant.REDName},
-                            {"Variant", varnt.Variant.ToObject()}
-                        };
-                    case IREDCurvePoint cp:
-                        if (cp.GetValue() is Tuple<IEditableVariable, IEditableVariable>(var item1, var item2))
-                        {
-                            return new Dictionary<string, object>() {{"Value", item1.ToObject()}, {"Point", item2.ToObject()}};
-                        }
-                        throw new InvalidParsingException($"{data.REDType} ToObject");
-                    // serialize primitive types directly
-                    case IREDPrimitive b:
-                        return b.GetValue();
-                    // serialize arrays as list of objects
-                    // serialize curves as array
-                    case IREDArray arr:
-                    {
-                        if (arr is IList ilist)
-                        {
-                            if (arr.IsByteArray())
-                            {
-                                return arr.GetBytes();
-                            }
-                            else
-                            {
-                                return ilist.Cast<IEditableVariable>().Select(_ => _.ToObject());
-                            }
-                        }
-
-                        throw new InvalidParsingException("Invalid File");
-                }
-                    case ICurveDataAccessor:
-                    {
-                        //TODO: bytearrays
-
-                        dynamic array = data;
-                        if (array.Elements is not IList dyn)
-                        {
-                            throw new InvalidParsingException("Invalid File");
-                        }
-                        return dyn
-                            .Cast<IEditableVariable>()
-                            .Select(_ => _.ToObject());
-                    }
-
-                    // serialize complex properties as Dictionary
-                    default:
-                        var dict = new Dictionary<string, object>();
-                        foreach (var cvar in data.SerializedProperties)
-                        {
-                            dict.Add(cvar.REDName, cvar.ToObject());
-                        }
-
-                        return dict;
-                }
-
-        }
-    }
-
-
-    #region models
-
-    public class Red4W2rcFileDto
-    {
-        public const string Magic = "w2rc";
-
-        public Dictionary<int,CR2WExportWrapperDto> Chunks { get; set; } = new();
-
-        public List<CR2WBufferWrapperDto> Buffers { get; set; } = new();
-
-        public bool ShouldSerializeBuffers() => Buffers is { Count: > 0 };
-
-        public Red4W2rcFileDto()
-        {
-
-        }
-
-        public Red4W2rcFileDto(IWolvenkitFile cr2w)
-        {
-            Chunks = cr2w.Chunks.ToDictionary(_ => _.ChunkIndex, _ => new CR2WExportWrapperDto(_));
-            Buffers = cr2w.Buffers.Select(_ => new CR2WBufferWrapperDto(_)).ToList();
-        }
-
-        public CR2WFile ToW2rc()
-        {
-            var cr2w = new CR2WFile
-            {
-                Buffers = Buffers
-                    .OrderBy(_ => _.Index)
-                    .Select(_ => _.ToRedBuffer())
-                    .ToList()
-            };
-
-            // chunks
-            // order so that parent chunks get created first
-            var groupedChunks = Chunks.GroupBy(_ => _.Value.ParentIndex);
-            foreach (IGrouping<int, KeyValuePair<int, CR2WExportWrapperDto>> groupedChunk in groupedChunks)
-            {
-                foreach (var (chunkIndex, chunk) in groupedChunk.OrderBy(_ => _.Key))
-                {
-                    chunk.CreateChunkInFile(cr2w, chunkIndex);
-                }
-            }
-
-            return cr2w;
-        }
-    }
-
-    public class CR2WExportWrapperDto
-    {
-        public string Type { get; set; }
-
-        //public int ChunkIndex { get; set; }
-        public int ParentIndex { get; set; }
-        public Dictionary<string, object> Properties { get; set; } = new();
-
-        public bool ShouldSerializeProperties() => Properties is { Count: > 0 };
-
-        public CR2WExportWrapperDto()
-        {
-            
-        }
-
-        public CR2WExportWrapperDto(ICR2WExport cr2WExport)
-        {
-            Type = cr2WExport.REDType;
-
-            //ChunkIndex = cr2WExport.ChunkIndex;
-            ParentIndex = cr2WExport.ParentChunkIndex;
-            var cvarAsDict = cr2WExport.Data.ToObject();
-            if (cvarAsDict is Dictionary<string, object> dict)
-            {
-                Properties = dict;
-            }
-            else
-            {
-                throw new SerializationException();
-            }
-        }
-
-        public void CreateChunkInFile(CR2WFile cr2WFile, int idx)
-        {
-            ICR2WExport parentChunk;
-            if (ParentIndex == -1)
-            {
-                parentChunk = null;
-            }
-            else
-            {
-                var foundChunk = cr2WFile.Chunks.FirstOrDefault(_ => _.ChunkIndex == ParentIndex);
-                if (foundChunk is null)
-                {
-                    throw new SerializationException("No parentchunk found");
-                }
-                else
-                {
-                    parentChunk = foundChunk;
-                }
-                
-            }
-
-            // create wrapped Cvariable
-            var cvar = CR2WTypeManager.Create(Type, Type, cr2WFile, parentChunk?.Data as CVariable);
-            // add data
-            cvar.SetFromDictionary(Properties);
-
-            cr2WFile.CreateChunk(cvar, idx, parentChunk as CR2WExportWrapper);
-        }
-
-    }
-
-    public class CR2WBufferWrapperDto
-    {
-        public uint Flags { get; set; }
-        public uint Index { get; set; }
-
-        // public uint Offset { get; set; }
-        // public uint DiskSize { get; set; }
-        // public uint MemSize { get; set; }
-        // public uint Crc32 { get; set; }
-
-        public CR2WBufferWrapperDto()
-        {
-
-        }
-
-        public CR2WBufferWrapperDto(ICR2WBuffer cr2WBuffer)
-        {
-            Flags = cr2WBuffer.Flags;
-            Index = cr2WBuffer.Index;
-            // Offset = cr2WBuffer.Offset;
-            // DiskSize = cr2WBuffer.DiskSize;
-            // MemSize = cr2WBuffer.MemSize;
-            // Crc32 = cr2WBuffer.Crc32;
-        }
-
-        public ICR2WBuffer ToRedBuffer() =>
-            new CR2WBufferWrapper(new CR2WBuffer()
-            {
-                flags = Flags,
-                index = Index,
-                // offset = Offset,
-                // diskSize = DiskSize,
-                // memSize = MemSize,
-                // crc32 = Crc32,
-            });
-    }
-
-    #endregion
-
-}
