+++ WolvenKit.Modkit/RED4/Tools/MaterialTools.cs
@@ -19,6 +19,7 @@ using WolvenKit.Common.Tools;
 using Newtonsoft.Json;
 using WolvenKit.Modkit.RED4.RigFile;
 using WolvenKit.Common;
+using WolvenKit.Common.Conversion;
 
 namespace WolvenKit.Modkit.RED4
 {
@@ -325,7 +326,7 @@ namespace WolvenKit.Modkit.RED4
                                     }
                                     var hp = _wolvenkitFileService.TryReadCr2WFile(ms);
                                     hp.FileName = primaryDependencies[i];
-                                    var dto = new Red4W2rcFileDto(hp);
+                                    var dto = new RedFileDto(hp);
                                     var doc = JsonConvert.SerializeObject(dto, settings);
                                     File.WriteAllText(path, doc);
                                 }
@@ -357,7 +358,7 @@ namespace WolvenKit.Modkit.RED4
                                         Directory.CreateDirectory(new FileInfo(path).Directory.FullName);
                                     }
                                     mls.FileName = primaryDependencies[i];
-                                    var dto = new Red4W2rcFileDto(mls);
+                                    var dto = new RedFileDto(mls);
                                     var doc = JsonConvert.SerializeObject(dto, settings);
                                     File.WriteAllText(path, doc);
                                 }
@@ -406,7 +407,7 @@ namespace WolvenKit.Modkit.RED4
                                                             Directory.CreateDirectory(new FileInfo(path1).Directory.FullName);
                                                         }
                                                         mlt.FileName = mls.Imports[e].DepotPathStr;
-                                                        var dto1 = new Red4W2rcFileDto(mlt);
+                                                        var dto1 = new RedFileDto(mlt);
                                                         var doc1 = JsonConvert.SerializeObject(dto1, settings);
                                                         File.WriteAllText(path1, doc1);
                                                     }
